requirements.txt
pandas
pyarrow
deltalake
kaggle
python-dotenv
